feat: ✨ 아이템 거래 생성/수정/삭제 API 개발
- Swagger UI 추가

+++ src/main/kotlin/com/tradecrossing/api/trade/VillagerTradeResource.kt
@@ -27,7 +27,7 @@ class VillagerTradeResource {
   class Create(val parent: VillagerTradeResource = VillagerTradeResource())
 
   @Resource("/{id}")
-  class Id(val parent: VillagerTradeResource = VillagerTradeResource(), val id: Int)
+  class Id(val parent: VillagerTradeResource = VillagerTradeResource(), val id: Long)
 
   companion object {
     val list: OpenApiRoute.() -> Unit = {
